requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.1
SQLAlchemy==2.0.32
Werkzeug==3.0.4
python-dotenv==1.0.1
requests==2.32.3
